+++ src/main/java/ii/cipriantarlev/marketmanagementapi/myorganization/MyOrganizationRepository.java
@@ -5,6 +5,6 @@ package ii.cipriantarlev.marketmanagementapi.myorganization;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MyOrganizationRepository extends JpaRepository<MyOrganization, Integer> {
+public interface MyOrganizationRepository extends JpaRepository<MyOrganization, Long> {
 
 }
