+++ src/main/java/ii/cipriantarlev/marketmanagementapi/subcategory/SubcategoryRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubcategoryRepository extends JpaRepository<Subcategory, Integer> {
+public interface SubcategoryRepository extends JpaRepository<Subcategory, Long> {
 
-	List<Subcategory> findAllByCategoryId(Integer categoryId);
+	List<Subcategory> findAllByCategoryId(Long categoryId);
 }
